+++ src/LumengPHP/Http/Result/Failed.php
@@ -3,7 +3,8 @@
 namespace LumengPHP\Http\Result;
 
 /**
- * 失败的结果<br />
+ * 失败的结果
+ * 
  * 当一个<b>Result</b>对象的status小于或等于0时，则表示失败
  *
  * @author zhengluming <luming.zheng@shandjj.com>
@@ -11,12 +12,13 @@ namespace LumengPHP\Http\Result;
 class Failed extends Result {
 
     /**
-     * 构造Failed对象
+     * 构造一个<b>Failed</b>实例
+     * 
      * @param string $msg 失败消息
-     * @param array $data (携带的)数据
+     * @param mixed $data (携带的)数据
      */
-    public function __construct($msg = '', array $data = []) {
-        parent::__construct(0, $msg, $data);
+    public function __construct($msg = '', $data = null) {
+        parent::__construct(self::FAILED, $msg, $data);
     }
 
 }