+++ src/LumengPHP/Http/Result/ResultHandlerInterface.php
@@ -12,22 +12,25 @@ use Exception;
 interface ResultHandlerInterface {
 
     /**
-     * 处理控制器方法返回结果，转换控制器方法返回结果为Result对象
+     * 处理控制器方法返回结果，转换控制器方法返回结果为ResultInterface对象
+     * 
      * @param mixed $return 控制器方法返回结果
-     * @return Result
+     * @return ResultInterface
      */
     public function handleReturn($return);
 
     /**
-     * 处理异常，转换异常对象为Result对象
+     * 处理异常，转换异常对象为ResultInterface对象
+     * 
      * @param Exception $ex 异常对象
-     * @return Result
+     * @return ResultInterface
      */
     public function handleException(Exception $ex);
 
     /**
      * 处理结果
-     * @param Result $result Result实例
+     * 
+     * @param ResultInterface $result ResultInterface实例
      */
-    public function handleResult(Result $result);
+    public function handleResult(ResultInterface $result);
 }