+++ src/LumengPHP/Http/Result/Success.php
@@ -3,7 +3,8 @@
 namespace LumengPHP\Http\Result;
 
 /**
- * 成功的结果<br />
+ * 成功的结果
+ * 
  * 当一个<b>Result</b>对象的status等于1时，则表示成功
  * 
  * @author zhengluming <luming.zheng@shandjj.com>
@@ -11,7 +12,8 @@ namespace LumengPHP\Http\Result;
 class Success extends Result {
 
     /**
-     * 构造Success对象
+     * 构造一个<b>Success</b>实例
+     * 
      * <pre>
      * Usage：
      * //场景1：只返回数据而不包含消息，可以直接返回一个array
@@ -23,11 +25,12 @@ class Success extends Result {
      * //场景3：返回成功消息及数据
      * return new Success('执行成功！', ['name' => '张三', 'age' => 18]);
      * </pre>
+     * 
      * @param string $msg 成功消息
-     * @param array $data (携带的)数据
+     * @param mixed $data (携带的)数据
      */
-    public function __construct($msg = '', array $data = []) {
-        parent::__construct(1, $msg, $data);
+    public function __construct($msg = '', $data = null) {
+        parent::__construct(self::SUCCESS, $msg, $data);
     }
 
 }